+++ ts/packages/knowPro/src/conversation.ts
@@ -2,14 +2,7 @@
 // Licensed under the MIT License.
 
 import { PromptSection } from "typechat";
-import {
-    IConversation,
-    DateRange,
-    SemanticRef,
-    SemanticRefOrdinal,
-    IMessage,
-    MessageOrdinal,
-} from "./interfaces.js";
+import { IConversation, DateRange } from "./interfaces.js";
 import { openai, TextEmbeddingModel } from "aiclient";
 import {
     TextEmbeddingIndexSettings,
@@ -17,7 +10,6 @@ import {
 } from "./fuzzyIndex.js";
 import { MessageTextIndexSettings } from "./messageIndex.js";
 import { RelatedTermIndexSettings } from "./relatedTermsIndex.js";
-import { Collection } from "./collections.js";
 
 export type ConversationSettings = {
     relatedTermIndexSettings: RelatedTermIndexSettings;
@@ -90,21 +82,4 @@ export function getTimeRangePromptSectionForConversation(
         ];
     }
     return [];
-}
-
-export class SemanticRefCollection extends Collection<
-    SemanticRef,
-    SemanticRefOrdinal
-> {
-    constructor(semanticRefs?: SemanticRef[]) {
-        super(semanticRefs);
-    }
-}
-
-export class MessageCollection<
-    TMessage extends IMessage = IMessage,
-> extends Collection<TMessage, MessageOrdinal> {
-    constructor(messages?: TMessage[]) {
-        super(messages);
-    }
 }
