+++ ts/packages/knowPro/src/index.ts
@@ -7,15 +7,17 @@ export * from "./conversation.js";
 export * from "./conversationIndex.js";
 export * from "./secondaryIndexes.js";
 export * from "./relatedTermsIndex.js";
-export * from "./search.js";
 export * from "./conversationThread.js";
 export * from "./fuzzyIndex.js";
 export * from "./propertyIndex.js";
 export * from "./timestampIndex.js";
-export * from "./serialization.js";
-export * from "./dateTimeSchema.js";
-export * from "./searchQueryTranslator.js";
 export * from "./textLocationIndex.js";
 export * from "./messageIndex.js";
+export * from "./search.js";
+export * from "./serialization.js";
 
 export * as querySchema from "./searchQuerySchema.js";
+export * from "./dateTimeSchema.js";
+export * from "./searchQueryTranslator.js";
+
+export * from "./storageProvider.js";
